+++ 3ETool_res/res_version.dat
@@ -1,1 +1,1 @@
-0.2.5
+0.3.0
