+++ EEETools/MainModules/support_blocks.py
@@ -1,5 +1,4 @@
 from xml.etree import ElementTree as ETree
-
 from EEETools.MainModules import Block, ArrayHandler
