+++ EEETools/Tools/API/terminal_api.py
@@ -1,13 +1,12 @@
+from EEETools.Tools.API.ExcelAPI.modules_importer import calculate_excel, import_excel_input
 from EEETools.Tools.GUIElements.connection_and_block_check import CheckConnectionWidget
-from EEETools.Tools.modules_importer import calculate_excel, import_excel_input
 from EEETools.Tools.GUIElements.net_plot_modules import display_network
 from EEETools.MainModules.main_module import CalculationOptions
-import os, requests, warnings
 from tkinter import filedialog
 from EEETools import costants
 from shutil import copyfile
-from github import Github
 import tkinter as tk
+import os, warnings
 
 
 def calculate(excel_path="", calculate_on_pf_diagram=True, loss_cost_is_zero=True, valve_is_dissipative=True,
@@ -82,13 +81,14 @@ def __import_file(filename):
         return
 
     file_path = os.path.join(dir_path, filename)
-    file_position = os.path.join(costants.RES_DIR, "Other", filename)
+    file_position = os.path.join(costants.RES_DIR, "../Other", filename)
 
     if not os.path.isfile(file_position):
 
         try:
 
-            __retrieve_file(filename, file_position)
+            from EEETools.Tools.Other.resource_downloader import update_resource_folder
+            update_resource_folder()
 
         except:
 
@@ -100,8 +100,6 @@ def __import_file(filename):
             warning_message += "new file position:\t" + file_path + "\n\n"
 
             warnings.warn(warning_message)
-
-            __retrieve_file(filename, file_position)
 
         else:
 
@@ -123,34 +121,3 @@ def __import_file(filename):
             warning_message += "file position:\t" + file_path + "\n\n"
 
             warnings.warn(warning_message)
-
-
-def __retrieve_file(filename, file_position):
-
-    token = costants.GITHUB_CONGIF["token"]
-    base_repo = costants.GITHUB_CONGIF["repo"]
-    base_path = costants.GITHUB_CONGIF["path"]
-
-    github = Github(token)
-    repo = github.get_user().get_repo(base_repo)
-    content = repo.get_contents(base_path + filename)
-
-    url = content.download_url
-    r = requests.get(url)
-
-    if r.status_code == 200:
-
-        with open(file_position, "wb") as file:
-
-            file.write(r.content)
-
-    else:
-
-        warning_message = "\n\n<----------------- !ERROR! ------------------->\n"
-        warning_message += "Unable to download the file!\n\n"
-
-        warning_message += "file name:\t\t" + filename + "\n"
-        warning_message += "status_code:\t\t" + str(r.status_code) + "\n"
-        warning_message += "url:\t\t\t" + url + "\n\n"
-
-        raise RuntimeError(warning_message)
