+++ EEETools/__init__.py
@@ -1,2 +1,2 @@
-from EEETools.Tools.main_tools import paste_default_excel_file, paste_components_documentation, paste_user_manual
-from EEETools.Tools.main_tools import calculate, launch_connection_debug, launch_network_display
+from EEETools.Tools.API.terminal_api import paste_default_excel_file, paste_components_documentation, paste_user_manual
+from EEETools.Tools.API.terminal_api import calculate, launch_connection_debug, launch_network_display
