+++ EEETools/version.py
@@ -1,1 +1,1 @@
-VERSION = "0.2.7"
+VERSION = "0.3.0"
